+++ src/Readarr.Api.V1/Wanted/UserFileUpgradeResource.cs
@@ -1,0 +1,13 @@
+using Readarr.Api.V1.Books;
+using Readarr.Http.REST;
+
+namespace Readarr.Api.V1.Wanted
+{
+    public class UserFileUpgradeResource : RestResource
+    {
+        public int BookId { get; set; }
+        public BookResource Book { get; set; }
+        public bool NeedsEpub { get; set; }
+        public bool NeedsM4b { get; set; }
+    }
+}